pyproject.toml
[project]
name = "opentimes"
version = "0.0.1"
description = "Free travel times between U.S. Census geographies"
keywords = [
  "accessibility",
  "spatial access",
  "travel times",
  "travel time estimation"
]
authors = [
  {name = "Dan Snow", email = "info@opentimes.org"}
]
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.3"
]

[project.urls]
repository = "https://github.com/dfsnow/opentimes"
homepage = "https://opentimes.org"

[project.optional-dependencies]
data = [
  "duckdb==1.1.2",
  "dvc[s3]==3.55.2",
  "geopandas[all]==1.0.1",
  "pandas==2.2.3",
  "pyarrow==17.0.0",
  "pyyaml==6.0.2",
  "requests==2.32.3"
]
site = [
  "boto3==1.35.35",
  "duckdb==1.1.2",
  "jinja2==3.0.3",
  "pyyaml==6.0.2",
  "requests==2.32.3"
]
dev = [
  "setuptools>=61.0",
  "pandas-stubs>=2.2.3.241009",
  "pre-commit>=4.0.1",
  "pytest>=7.3.0",
  "pytest-cov>=4.1.0",
  "types-PyYAML>=6.0.12.20240917",
  "types-requests>=2.32.0.20241016"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["opentimes"]

[tool.pyright]
# Let ruff handle these
reportUnusedImport = false
reportUndefinedVariable = false

[tool.ruff]
line-length = 79

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
